WIP: Rust <-> JS bindings
