+++ noise/Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["host", "methods", "noise", "map"]
+members = ["host", "methods", "noise", "map", "js"]
 
 # Always optimize; building and running the guest takes much longer without optimization.
 [profile.dev]
