+++ noise/js/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "gb-noise"
+version = "0.2.0"
+authors = ["Dzejkop <jakubtrad@gmail.com>"]
+edition = "2018"
+description = "Gemstone Builders noise"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[features]
+default = ["console_error_panic_hook"]
+
+[dependencies]
+wasm-bindgen = "0.2.84"
+js-sys = "0.3.69"
+noise = { path = "../noise", default-features = false }
+fixed = { version = "1.27.0", features = ["serde-str"] }
+
+# The `console_error_panic_hook` crate provides better debugging of panics by
+# logging them with `console.error`. This is great for development, but requires
+# all the `std::fmt` and `std::panicking` infrastructure, so isn't great for
+# code size when deploying.
+console_error_panic_hook = { version = "0.1.7", optional = true }
+
+[dev-dependencies]
+wasm-bindgen-test = "0.3.34"
