+++ noise/js/README.md
@@ -1,0 +1,1 @@
+# Gemstone Builders Noise package
